+++ Core/Models/Startup.cs
@@ -27,7 +27,7 @@ namespace Core.Models
         public long? Revenue { get; set; }
         public int? DAU { get; set; }
         public int? MAU { get; set; }
-        public int? GrowthPercentage { get; set; }
+        public double? GrowthPercentage { get; set; }
 
         public string PitchDeckUrl { get; set; }
         public string FinancialModelUrl { get; set; }
